Restructure bannerbutton rendering
Adds user ranks and rank-based permissions, as well as methods to the Project model and controller to make retrieving lists of projects easier.

+++ app/Http/Controllers/BlogPostController.php
@@ -6,30 +6,29 @@ use App\Models\BlogPost;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Redirect;
+use App\Models\User;
 
 class BlogPostController extends Controller
 {
     public function index()
     {
         $posts = BlogPost::all();
+        foreach($posts as $post) {
+            $post->user_name = User::where("id", $post->id)->firstOrFail()->name;
+        }
         return view( "blog.index", ["posts" => $posts] );
     }
 
     public function create()
     {
-        if(!Auth::check()) {
-            return redirect("/login?return=/blog/create");
-        }
+        if(!Auth::check()) return route("login");
         return view("blog.create");
     }
 
 
     public function store(Request $request)
     {
-        if(!Auth::check()) {
-            return redirect("/login?return=/blog/create");
-        }
-
+        if(!Auth::check()) return route("login");
         $newPost = BlogPost::create(
             [
                 "title" => $request->title,
@@ -48,18 +47,20 @@ class BlogPostController extends Controller
 
     public function edit(BlogPost $blogPost)
     {
-        if(Auth::id() != $blogPost->user_id) {
+        if(!Auth::check()) return route("login");
+        if(!$blogPost->edit_allowed()) {
             return Redirect::back()->withErrors("You do not have permission to edit this post.");
         }
+        //$blogPost->user_name = $blogPost->findCreator();
         return view("blog.edit", ["blogPost" => $blogPost]);
     }
 
     public function update(Request $request, BlogPost $blogPost)
     {
-        if($request->user()->id != $blogPost->user_id) {
+        if(!Auth::check()) return route("login");
+        if(!$blogPost->edit_allowed()) {
             return Redirect::back()->withErrors("You do not have permission to edit this post.");
         }
-
         $blogPost->update(
             [
                 "title" => $request->title,
@@ -71,6 +72,9 @@ class BlogPostController extends Controller
 
     public function destroy(BlogPost $blogPost)
     {
+        if(!$blogPost->edit_allowed()) {
+            return Redirect::back()->withErrors("You do not have permission to edit this post.");
+        }
         $blogPost->delete();
         return redirect("/blog");
     }
